Put `explcheck-preprocessing-comments` to `parsers` and `preprocessing`
