+++ explcheck/src/explcheck-parsers.lua
@@ -1,7 +1,7 @@
 -- Common LPEG parsers used by different modules of the static analyzer explcheck.
 
 local lpeg = require("lpeg")
-local P, R, S = lpeg.P, lpeg.R, lpeg.S
+local Cp, P, R, S = lpeg.Cp, lpeg.P, lpeg.R, lpeg.S
 
 -- Base parsers
 ---- Generic
@@ -16,6 +16,7 @@ local backslash = P([[\]])
 local letter = R("AZ","az")
 local underscore = P("_")
 local colon = P(":")
+local percent_sign = P("%")
 
 ---- Spacing
 local newline = (
@@ -33,8 +34,9 @@ local optional_spaces_and_newline = (
     * optional_spaces
   )^-1
 )
+local blank_line = optional_spaces * newline
 
--- Intermediate parsers.
+-- Intermediate parsers
 ---- Parts of TeX syntax
 local argument = (
   lbrace
@@ -70,6 +72,54 @@ local expl3like_material = (
   + expl3_variable_or_constant
 )
 
+local commented_line_letter = (
+  linechar
+  + newline
+  - backslash
+  - percent_sign
+)
+local commented_line = (
+  (
+    (
+      commented_line_letter
+      - newline
+    )^1  -- initial state
+    + (
+      backslash  -- even backslash
+      * (
+        backslash
+        + #newline
+      )
+    )^1
+    + (
+      backslash
+      * (
+        percent_sign
+        + commented_line_letter
+      )
+    )
+  )^0
+  * (
+    #percent_sign
+    * Cp()
+    * (
+      (
+        percent_sign  -- comment
+        * linechar^0
+        * Cp()
+        * newline
+        * #blank_line  -- blank line
+      )
+      + percent_sign  -- comment
+      * linechar^0
+      * Cp()
+      * newline
+      * optional_spaces  -- leading spaces
+    )
+    + newline
+  )
+)
+
 ---- Standard delimiters
 local provides = (
   P([[\ProvidesExpl]])
@@ -92,6 +142,7 @@ local expl_syntax_off = P([[\ExplSyntaxOff]])
 
 return {
   any = any,
+  commented_line = commented_line,
   eof = eof,
   expl3like_material = expl3like_material,
   expl_syntax_off = expl_syntax_off,
