+++ explcheck/src/explcheck-preprocessing.lua
@@ -1,11 +1,44 @@
 -- The preprocessing step of static analysis determines which parts of the input files contain expl3 code.
 
 local parsers = require("explcheck-parsers")
-local strip_comments = require("explcheck-preprocessing-comments")
 local utils = require("explcheck-utils")
 
 local lpeg = require("lpeg")
-local Cp, P, V = lpeg.Cp, lpeg.P, lpeg.V
+local Cp, Ct, P, V = lpeg.Cp, lpeg.Ct, lpeg.P, lpeg.V
+
+-- Strip TeX comments from a text. Besides the transformed text, also return
+-- a function that maps positions in the transformed text back to the original
+-- text.
+local function strip_comments(text)
+  local transformed_index = 0
+  local numbers_of_bytes_removed = {}
+  local transformed_text_table = {}
+  for index, text_position in ipairs(lpeg.match(Ct(parsers.commented_line^1), text)) do
+    local span_size = text_position - transformed_index - 1
+    if span_size > 0 then
+      if index % 2 == 1 then  -- chunk of text
+        table.insert(transformed_text_table, text:sub(transformed_index + 1, text_position - 1))
+      else  -- comment
+        table.insert(numbers_of_bytes_removed, {transformed_index, span_size})
+      end
+      transformed_index = transformed_index + span_size
+    end
+  end
+  table.insert(transformed_text_table, text:sub(transformed_index + 1, -1))
+  local transformed_text = table.concat(transformed_text_table, "")
+  local function map_back(index)
+    for _, where_and_number_of_bytes_removed in ipairs(numbers_of_bytes_removed) do
+      local where, number_of_bytes_removed = table.unpack(where_and_number_of_bytes_removed)
+      if index > where then
+        index = index + number_of_bytes_removed
+      else
+        break
+      end
+    end
+    return index
+  end
+  return transformed_text, map_back
+end
 
 -- Preprocess the content and register any issues.
 local function preprocessing(issues, content, options)
